Pull out Dijkstra on HandleGraphs into its own function
